backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0
alembic>=1.12.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP and Requests
httpx>=0.25.0
aiohttp>=3.9.0

# Validation and Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Environment and Configuration
python-dotenv>=1.0.0

# Caching
redis>=5.0.0

# Voice Processing
openai>=1.3.0

# Memory Management
zep-python>=2.0.0

# Vector Database
pinecone-client>=2.2.0

# Text-to-Speech
elevenlabs>=0.2.0

# Content Processing
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
python-slugify>=8.0.0
python-dateutil>=2.8.0